+++ src/main/java/searchengine/repositories/IndexRepository.java
@@ -11,7 +11,7 @@ import searchengine.model.Page;
 import java.util.List;
 
 @Repository
-public interface IndexRepository extends JpaRepository<ModelIndex, Long> {
+public interface IndexRepository extends JpaRepository<ModelIndex, Integer> {
     @Query(value = "SELECT i.* FROM `index` i WHERE i.lemma_id IN :lemmas AND i.page_id IN :pages", nativeQuery = true)
     List<ModelIndex> findByPagesAndLemmas(@Param("lemmas") List<Lemma> lemmaListId, @Param("pages") List<Page> pageListId);
 }
