+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -12,7 +12,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface PageRepository extends JpaRepository<Page, Long> {
+public interface PageRepository extends JpaRepository<Page, Integer> {
     long countBySiteId(SitePage siteId);
     Iterable<Page> findBySiteId(SitePage site);
     @Query(value = "SELECT p.* FROM Page p JOIN `index` i ON p.id = i.page_id WHERE i.lemma_id IN :lemmas", nativeQuery = true)
