+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import searchengine.model.SitePage;
 @Repository
-public interface SiteRepository extends JpaRepository<SitePage, Long> {
+public interface SiteRepository extends JpaRepository<SitePage, Integer> {
     SitePage findByUrl(String url);
-    SitePage findByUrl(long id);
+    SitePage findByUrl(int id);
     SitePage findByUrl(SitePage site);
 }
